Multiplayer Tweaking
Prepared game for playtesting, including some debug variables for
skipping parts of the game (in order to streamline gameplay).
Fixed several issues related to actually compiling the game into a
runnable jar.
Attempts were made to try to improve syncing of player positions in
multiplayer.
Player robots are no longer several parts welded together; instead its
entire collision is combined into one body (this fixes the issue of the
welds freaking out and corrupting the robot).
ClawGrabberComponent now welds the grabbed power cube to itself when
grabbed, and destroys the weld when released (this change was needed
because dynamically changing the collision of the robot is inefficient
due to the above change with robot collision).