Make Packets Synchronized
Instead of running packets as soon as they are recieved, both the client
and server queue recieved packets, which are ran the next tick. This
keeps the packets synchronized with the game world to avoid concurrent
modification.
This appears to have fixed the problem with the robots freaking out and
teleporting everywhere as well as the problem of the server hanging
during world updates.