Fix Cube Syncing + Sync Switch, Score, Height
Picking up and dropping cubes works properly again. Cubes no longer
count as on a platform if they're being held.
Score is synchronized with the server.
Which side of the switches and scales belong to what team (orientation)
is now synchronized with the server.
Lifter height (and climbing) is synchronized with the server.
The robots' positions at the start of the match are synchronized with
the server.
The server's window now fits to the aspect ratio of the field.