More Syncing + Bug Fixes
Game time is now syncronized with the server.
Players' teams are now synchronized with the server.
When a player joins, their team is set. Currently, the team chosen
alternates between blue and red.
If the client recieves a CubeUpdatePacket for a cube that doesn't exist,
it creates the cube.
Fixed a crash on connecting to the server.