Synchronized Ejecting Cubes from Portals/Exchange
Ejecting cubes onto the field from the portals or exchange is now
synchronized with the server.
The portals and exchanges now work for both teams.
The HUD's sprites for the portals and exchange now reflect which team
you're on.
Fixed the blue team not starting with the correct number of cubes
stored.