Make the Client UI Different if there's no Server
The client now doesn't look like it's connected to a server if it isn't
connected to a server.
The client UI now doesn't say it's waiting for players if it isn't
connected to a server.
Voting to start if the client is unconnected just force starts, so the
UI was updated to reflect this.