Fix Some Issues During First Playtesting
Instead of the server sending a joining client everyone else's robots
after a delay, the client now requests everyone's robot after it's ready
to recieve them. This prevents the issue where laggy connections would
cause the clients not to recieve the robot data packets in time.
When in the debug view (numpad 4), the ui now shows a list of all
players, as well as if they have a robot and, for the server, the
hashcode of the Connection corresponding to the player.
The Player constructor used to block while attempting to poll
TheBlueAlliance; this is done in another Thread now.
Removed ShipComponentHealthPacket because it's not needed.
Lowered the sleep time in WorldUpdateThread for the server Thread so
that the server physics better align with the clients' physics (this has
the benefit of making the power cube physics much more tolerable for the
clients).