Handle Hanging of the Game Loop Thread
The game loop now runs in its own Thread instead of the main Thread.
There is now a monitor Thread that checks if the game loop thread is
hung. If it finds the game loop to be hung, it forcefully terminates it,
and starts a new thread in its place.
This will hopefully solve the problems with the main thread hanging
until the actual problem is found.
There is a button for triggering a debug hang; holding shift and
pressing the key instead toggles debug lag.