Add Command-Line Arguments
The client can optionally take an ip-port combo as an argument now (x
"127.0.0.1:50001"). It will automatically connect if this argument is
provided. If it isn't provided, the client will default to prompting the
user.
The server can optionally take the port as an argument now. If it isn't
provided, the server will use the default port (50001).